+++ RELEASE_NOTES.md
@@ -3,7 +3,8 @@ Release Notes
 ## 1.2.0-beta2
 * Container Groups: Support for image registry credentials for private registries.
 * Container Groups: Support for partial CPU cores.
-* Internal changes to better capture full resource IDs
+* Log Analytics: Initial support.
+* Internal changes to better capture full resource IDs.
 
 ## 1.2.0-beta1
 * Bastion Hosts: Create bastion hosts for accessing resources on a virtual network