+++ RELEASE_NOTES.md
@@ -2,9 +2,12 @@ Release Notes
 =============
 ## 1.2.0-beta2
 * Managed Identity: Create user assigned manage identity resources.
-* Container Groups: Specify identity settings to support user or system assigned identity.
 * Container Service (AKS): Specify identity settings to support user or system assigned identity.
-* Internal changes to better capture full resource IDs
+* Container Groups: Specify identity settings to support user or system assigned identity.
+* Container Groups: Support for image registry credentials for private registries.
+* Container Groups: Support for partial CPU cores.
+* Log Analytics: Initial support.
+* Internal changes to better capture full resource IDs.
 
 ## 1.2.0-beta1
 * Bastion Hosts: Create bastion hosts for accessing resources on a virtual network